+++ tests/test.hs
@@ -2,9 +2,25 @@
 {-#LANGUAGE FlexibleContexts#-}
 {-#LANGUAGE TemplateHaskell#-}
 {-#LANGUAGE QuasiQuotes#-}
+{-#LANGUAGE ScopedTypeVariables#-}
 
 import Test.Hspec
 import Language.Cube
+import Graphics.Formats.STL
+
+-- | Generate house which is for demo.
+house :: Block Cube
+house = let house''  = (house' + square)*line
+        in rfmap (12 * dz +) house''
+  where
+    house' :: Block Cube
+    house' = block $ [cube 1 x y| x<-[-10..10], y<-[-10..10], y < x , y < (-x)]
+    
+    square :: Block Cube
+    square = rfmap ((+) (-12 * dz)) $ block $ [cube 1 x y| x<-[-5..5], y<-[-5..5]]
+    
+    line :: Block Cube
+    line = block $ [cube x 0 0 | x<-[-5..5]]
 
 main :: IO ()
 main = 
@@ -23,26 +39,33 @@ main =
       it "-2 * dx" $ do
         (-2) * dx `shouldBe` cube (-2) 0 0 
       it "route pi/2" $ do
-        (dr (pi/2) dz) * dx `shouldBe` cube 0 1 0
+        (dR (pi/2) dz dx) `shouldBe` cube 0 1 0
       it "route pi" $ do
-        (dr (pi) dz) * dx `shouldBe` cube (-1) 0 0
+        (dR (pi) dz dx) `shouldBe` cube (-1) 0 0
       it "route 3*pi/2" $ do
-        (dr (3*pi/2) dz) * dx `shouldBe` cube 0 (-1) 0
+        (dR (3*pi/2) dz dx) `shouldBe` cube 0 (-1) 0
       it "route 2*pi" $ do
-        (dr (2*pi) dz) * dx `shouldBe` cube 1 0 0
+        (dR (2*pi) dz dx) `shouldBe` cube 1 0 0
       it "multi" $ do
-        dx * dy `shouldBe` dz
+        dx * dy `shouldBe` (dz :: Cube)
       it "multi" $ do
-        dy * dx `shouldBe` (-dz)
+        dy * dx `shouldBe` (-dz :: Cube)
     describe "Block" $ do
       it "dx" $ do
-        block [dx,dx] `shouldBe` block [dx]
+        block [dx,dx] `shouldBe` (block [dx] :: Block Cube)
       it "dx+dy" $ do
-        block [dx] + block [dy] `shouldBe` block [dx,dy]
+        block [dx] + block [dy] `shouldBe` (block [dx,dy] :: Block Cube)
       it "[dx,dy]-dy" $ do
-        block [dx,dy] - block [dy] `shouldBe` block [dx]
+        block [dx,dy] - block [dy] `shouldBe` (block [dx] :: Block Cube)
       it "convolution(dx * [-2*dx,2*dx])" $ do
-        block [dx] * block [-2*dx,2*dx] `shouldBe` block [-dx,3*dx]
+        block [dx] * block [-2*dx,2*dx] `shouldBe` (block [-dx,3*dx] :: Block Cube)
+    describe "Compaction" $ do
+      it "num triangle of one block" $ do
+        length (triangles (toSTL (block [dx] :: Block Cube))) `shouldBe` 12
+      it "num triangle of two block" $ do
+        length (triangles (toSTL (block [dx,2*dx] :: Block Cube))) `shouldBe` 20
+      it "flip triangle" $ do
+        (flipTriangle (Triangle Nothing ((5.0,0,0),(1,0,0),(2,0,0)))) `shouldBe` (Triangle Nothing ((5.0,0,0),(2,0,0),(1,0,0)))
     describe "ToSTL" $ do
       it "Block Cube" $ do
         writeFileStl "tmp.stl" (block [dx] :: Block Cube) `shouldReturn` ()