+++ ChangeLog.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+* Fix bug of rotation
+* Add RFunctor to define Functor for Data.Set and other data.
+* Add sample in cabal
+
 ## 0.1.0
 
 * First Release